E_032_Create_tests_for_BMICalculator_refactored
