+++ pom.xml
@@ -80,12 +80,6 @@
             <artifactId>modelmapper</artifactId>
             <version>2.3.9</version>
         </dependency>
-        <dependency>
-            <groupId>org.assertj</groupId>
-            <artifactId>assertj-core</artifactId>
-            <version>3.18.0</version>
-            <scope>test</scope>
-        </dependency>
 
     </dependencies>
